# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API handles Patients and Dentists.",
        "title": "Certified Tech Developer - Back End III",
        "termsOfService": "https://developers.ctd.com.ar/es_ar/terminos-y-condiciones",
        "contact": {
            "name": "API Support",
            "url": "https://developers.ctd.com.ar/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/odontologos/create": {
            "post": {
                "description": "Create a new odontologo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontologo"
                ],
                "summary": "Create a new odontologo",
                "parameters": [
                    {
                        "description": "Odontologo object",
                        "name": "odontologo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    }
                }
            }
        },
        "/odontologos/delete/{id}": {
            "delete": {
                "description": "Delete an odontologo by ID",
                "tags": [
                    "Odontologo"
                ],
                "summary": "Delete an odontologo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Odontologo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Odontologo deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/odontologos/update/{id}": {
            "put": {
                "description": "Update an existing odontologo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontologo"
                ],
                "summary": "Update an existing odontologo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Odontologo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Odontologo object",
                        "name": "odontologo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    }
                }
            }
        },
        "/odontologos/updateField/{id}": {
            "patch": {
                "description": "Update a specific field of an odontologo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontologo"
                ],
                "summary": "Update a specific field of an odontologo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Odontologo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Field name",
                        "name": "field_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "New value",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Field updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/odontologos/{id}": {
            "get": {
                "description": "Retrieve an odontologo by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Odontologo"
                ],
                "summary": "Get an odontologo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Odontologo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Odontologo"
                        }
                    }
                }
            }
        },
...
        "/pacientes/create": {
            "post": {
                "description": "Create a new patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Create a new patient",
                "parameters": [
                    {
                        "description": "Patient object",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                }
            }
        },
        "/pacientes/delete/{id}": {
            "delete": {
                "description": "Delete a patient by ID",
                "tags": [
                    "Paciente"
                ],
                "summary": "Delete a patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pacientes/update/{id}": {
            "put": {
                "description": "Update an existing patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Update an existing patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient object",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                }
            }
        },
        "/pacientes/updateField/{id}": {
            "patch": {
                "description": "Update a specific field of a patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Update a specific field of a patient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Field name",
                        "name": "field_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "New value",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "TOKEN",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Field updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pacientes/{id}": {
            "get": {
                "description": "Retrieve a patient by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Paciente"
                ],
                "summary": "Get a patient by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Paciente"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "domain.Odontologo": {
            "type": "object",
            "properties": {
                "apellido": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "domain.Paciente": {
            "type": "object",
            "properties": {
                "apellido": {
                    "type": "string",
                    "example": "Doe"
                },
                "dni": {
                    "type": "string",
                    "example": "12345678"
                },
                "domicilio": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "fecha_alta": {
                    "$ref": "#/definitions/domain.CustomTime"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nombre": {
                    "type": "string",
                    "example": "John"
                }
            }
        },
...
    }
}